Remove driver_test.go file and consolidate test functions into driver_with_context_test.go
